# fairway360/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fairway360 multi-tenant SaaS API. All data is scoped to the authenticated user's club; club_id is always derived from the session, never from the request.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [members]
      responses:
        "200":
          description: Club events
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ClubEvent" }
    post:
      operationId: createEvent
      tags: [supervisor]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EventInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ClubEvent" }
...
  /tasks:
    get:
      operationId: listTasks
      tags: [supervisor]
      responses:
        "200":
          description: All team tasks
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Task" }
    post:
      operationId: createTask
      tags: [supervisor]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TaskInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
...
  /time-off/{id}:
    patch:
      operationId: decideTimeOff
      tags: [supervisor]
      parameters: [{ $ref: "#/components/parameters/IdPath" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TimeOffDecision" }
      responses:
        "200":
          description: Decided
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TimeOff" }
...
  /requests/{id}/resolve:
    patch:
      operationId: resolveRequest
      tags: [supervisor]
      parameters: [{ $ref: "#/components/parameters/IdPath" }]
      responses:
        "200":
          description: Resolved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
  /staff:
    get:
      operationId: listStaff
      tags: [supervisor]
      responses:
        "200":
          description: Staff roster
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TeamMember" }
    post:
      operationId: createStaff
      tags: [supervisor]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateStaffInput" }
      responses:
        "201":
          description: Staff member created + invited
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InviteResult" }
        "400": { $ref: "#/components/responses/Error" }
  /members:
    get:
      operationId: listClubMembers
      tags: [supervisor]
      responses:
        "200":
          description: Club member roster
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ClubMember" }
    post:
      operationId: createMember
      tags: [supervisor]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateMemberInput" }
      responses:
        "201":
          description: Member created + invited
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InviteResult" }
        "400": { $ref: "#/components/responses/Error" }
  /staff/{id}/message:
    post:
      operationId: messageStaff
      tags: [supervisor]
      parameters: [{ $ref: "#/components/parameters/IdPath" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StaffMessageBody" }
      responses:
        "200":
          description: Message result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageResult" }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
...
  /leads/{id}:
    patch:
      operationId: updateLead
      tags: [supervisor]
      parameters: [{ $ref: "#/components/parameters/IdPath" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LeadUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Lead" }
...
  /bookings/{id}:
    patch:
      operationId: updateBooking
      tags: [supervisor]
      parameters: [{ $ref: "#/components/parameters/IdPath" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BookingUpdate" }
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
        "404": { $ref: "#/components/responses/Error" }
  /announcements:
    get:
      operationId: listAnnouncements
      tags: [members]
      responses:
        "200":
          description: Announcements
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Announcement" }
    post:
      operationId: createAnnouncement
      tags: [supervisor]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnnouncementInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Announcement" }

  # ---- Department channels (chat) -------------------------------------------
...
  /escalations/{id}/resolve:
    patch:
      operationId: resolveEscalation
      tags: [supervisor]
      parameters: [{ $ref: "#/components/parameters/IdPath" }]
      responses:
        "200":
          description: Resolved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
        "404": { $ref: "#/components/responses/Error" }

  # ---- Supervisor delegation ------------------------------------------------
...
  /agents/delegate:
    post:
      operationId: startDelegation
      tags: [supervisor]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DelegateInput" }
      responses:
        "200":
          description: Delegation started
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Delegation" }
...
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string }
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ProblemDetails" }
  schemas:
...
    Ok:
      type: object
      properties:
        ok: { type: boolean }
      required: [ok]
...
    Delegation:
      type: object
      properties:
        active: { type: boolean }
        autonomy: { type: string }
        by: { type: string }
        until: { type: string }
        startedAt: { type: string }
      required: [active]
    DelegateInput:
      type: object
      properties:
        autonomy:
          type: string
          enum: [low, medium, high, full]
        durationMinutes: { type: number }
      required: [autonomy, durationMinutes]
    CreateStaffInput:
      type: object
      properties:
        name: { type: string }
        email: { type: string }
        role: { type: string, enum: [employee, supervisor] }
        jobTitle: { type: string }
        phone: { type: string }
      required: [name, email, role, jobTitle]
    CreateMemberInput:
      type: object
      properties:
        name: { type: string }
        email: { type: string }
        tier: { type: string }
        phone: { type: string }
      required: [name, email]
    InviteResult:
      type: object
      properties:
        id: { type: string }
        inviteLink: { type: string }
        emailed: { type: boolean }
      required: [id, inviteLink, emailed]
    ClubMember:
      type: object
      properties:
        id: { type: string }
        userId: { type: string }
        name: { type: string }
        email: { type: string }
        tier: { type: string }
        memberNumber: { type: string }
        status: { type: string }
        pending: { type: boolean }
      required: [id, userId, name, email, tier, memberNumber, status, pending]
...
    ProblemDetails:
      type: object
      properties:
        title: { type: string }
        detail: { type: string }
        status: { type: number }
      required: [title]
...
    StaffMessageBody:
      type: object
      properties:
        message: { type: string, minLength: 1 }
      required: [message]
    MessageResult:
      type: object
      properties:
        sent: { type: boolean }
      required: [sent]
...
    ClubEvent:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        date: { type: string }
        time: { type: string }
        spots: { type: string }
        tag: { type: string }
      required: [id, title, date, time, spots, tag]
    EventInput:
      type: object
      properties:
        title: { type: string }
        startsAt: { type: string }
        tag: { type: string }
        capacity: { type: number }
      required: [title, startsAt, tag]
...
    Task:
      type: object
      properties:
        id: { type: string }
        label: { type: string }
        due: { type: string }
        done: { type: boolean }
        priority: { type: string }
        assignee: { type: string }
      required: [id, label, due, done]
    TaskInput:
      type: object
      properties:
        label: { type: string }
        dueAt: { type: string }
        priority: { type: string }
        assignedTo: { type: string }
      required: [label]
...
    TimeOffStatus:
      type: string
      enum: [Pending, Approved, Denied]
    TimeOff:
      type: object
      properties:
        id: { type: string }
        dates: { type: string }
        reason: { type: string }
        status: { $ref: "#/components/schemas/TimeOffStatus" }
        submitted: { type: string }
      required: [id, dates, status, submitted]
...
    TimeOffDecision:
      type: object
      properties:
        status: { $ref: "#/components/schemas/TimeOffStatus" }
      required: [status]
    TeamMember:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        initials: { type: string }
        role: { type: string }
        status: { type: string }
        shift: { type: string }
        area: { type: string }
        tasksOpen: { type: number }
      required: [id, name, initials, role, status]
...
    LeadStatus:
      type: string
      enum: [New, Contacted, Tour Booked, Won, Lost]
    Lead:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        contactName: { type: string }
        email: { type: string }
        phone: { type: string }
        source: { type: string }
        interest: { type: string }
        status: { $ref: "#/components/schemas/LeadStatus" }
        businessType: { type: string }
        problem: { type: string }
        volume: { type: string }
        time: { type: string }
      required: [id, name, status]
    LeadUpdate:
      type: object
      properties:
        status: { $ref: "#/components/schemas/LeadStatus" }
      required: [status]
    Booking:
      type: object
      properties:
        id: { type: string }
        member: { type: string }
        time: { type: string }
        players: { type: number }
        holes: { type: number }
        status: { type: string }
      required: [id, member, time, players, holes, status]
    BookingUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [confirmed, checked_in, cancelled]
      required: [status]
    Announcement:
      type: object
      properties:
        id: { type: string }
        tag: { type: string }
        title: { type: string }
        date: { type: string }
      required: [id, title]
    AnnouncementInput:
      type: object
      properties:
        tag: { type: string }
        title: { type: string }
        body: { type: string }
        audience: { type: string }
      required: [title]
...